pyproject.toml
[project]
name = "pilotdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.3.1",
    "flask>=3.1.1",
    "gdown>=5.2.0",
    "numpy>=2.3.1",
    "pandas>=2.3.0",
    "psycopg2-binary>=2.9.10",
    "pyodbc>=5.2.0",
    "pyyaml>=6.0.2",
    "scipy>=1.15.3",
    "sqlglot>=26.30.0",
]

[dependency-groups]
dev = [
    "pilotdb",
]

[tool.uv.sources]
pilotdb = { workspace = true }
